requirement.txt
python-dotenv
telebot
elevenlabs